// Views/src/components/financial/components/variable-expenses/components/list/dto/list-variable-expenses-dto.ts
import { ItemsInterface } from "../../../../../../../shared/components/list-g/list/interfaces/items-interface";


export class ListVariableExpensesDto {
  id: ItemsInterface;
  name: ItemsInterface;
  place:ItemsInterface;
  category: ItemsInterface;
  subcategory: ItemsInterface;
  paidDay: ItemsInterface;
  paidDayToView: ItemsInterface;
  price: ItemsInterface;
}